+++ packages/@baldeweg/ui/src/components/BFieldset/BFieldset.vue
@@ -7,7 +7,7 @@ defineProps({
 </script>
 
 <template>
-  <fieldset class="fieldset">
+  <fieldset class="fieldset u:clear-both u:mb-xl u:overflow-auto">
     <legend class="fieldset_legend">
       {{ legend }}
     </legend>
@@ -16,15 +16,6 @@ defineProps({
 </template>
 
 <style>
-.fieldset {
-  clear: both;
-  border: 0 hidden;
-  padding: 0;
-  margin: 0;
-  margin-bottom: 20px;
-  overflow: auto;
-}
-
 .fieldset:last-child {
   margin-bottom: 0;
 }