+++ packages/@baldeweg/ui/src/components/index.ts
@@ -1,7 +1,5 @@
 import BAlert from './BAlert/BAlert.vue';
 import BApp from './BApp/BApp.vue';
-import BBottomNav from './BBottomNav/BBottomNav.vue';
-import BBottomNavItem from './BBottomNav/BBottomNavItem.vue';
 import BButton from './BButton/BButton.vue';
 import BCard from './BCard/BCard.vue';
 import BCode from './BCode/BCode.vue';
@@ -14,11 +12,8 @@ import BFieldset from './BFieldset/BFieldset.vue';
 import BForm from './BForm/BForm.vue';
 import BInput from './BInput/BInput.vue';
 import BList from './BList/BList.vue';
-import BMasthead from './BMasthead/BMasthead.vue';
-import BMastheadItem from './BMasthead/BMastheadItem.vue';
 import BMaterialIcon from './BMaterialIcon/BMaterialIcon.vue';
 import BModal from './BModal/BModal.vue';
-import BNav from './BNav/BNav.vue';
 import BNavigation from './BNavigation/BNavigation.vue';
 import BNavigationItem from './BNavigation/BNavigationItem.vue';
 import BPanel from './BPanel/BPanel.vue';
@@ -36,8 +31,6 @@ import BUpload from './BUpload/BUpload.vue';
 export {
   BAlert,
   BApp,
-  BBottomNav,
-  BBottomNavItem,
   BButton,
   BCard,
   BCode,
@@ -50,11 +43,8 @@ export {
   BForm,
   BInput,
   BList,
-  BMasthead,
-  BMastheadItem,
   BMaterialIcon,
   BModal,
-  BNav,
   BNavigation,
   BNavigationItem,
   BPanel,