+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## v0.20.0
 
-- Removed `BTabs`, use `BNavigation` instead
+- Added Tailwind utility classes
+- Removed class `hidden`, use `u:hidden` instead
+- Removed class `visuallyHidden`, use `u:sr-only` instead
+- Removed `BBottomNav` component, use `BNavigation` instead
+- Removed `BNav` component, use `BNavigation` instead
+- Removed `BMasthead` component, no replacement
+- Removed `BTabs` component, use `BNavigation` instead
 
 ## v0.19.0